requirements.txt
easydict
einops
ftfy
huggingface-hub
jinja2
matplotlib
networkx
ninja
nltk
numpy
open3d
opencv-python
pandas
pillow
plyfile
pyyaml
regex
scikit-learn
scipy
timm
torch-geometric=2.3.1
torch-cluster=1.6.1
torch-scatter=2.1.1
torch-sparse=0.6.17
torch-spline-conv=1.2.2
tqdm
transformers